split javaUtils
