+++ src/commands/deploy/runPreDeployTask.ts
@@ -10,15 +10,16 @@ import * as vscode from 'vscode';
 import * as appservice from "vscode-azureappservice";
 import { localize } from '../../localize';
 import { javaUtils } from "../../utils/javaUtils";
+import { tryGetDeployableArtifacts } from "./java/tryGetDeployableArtifacts";
 
 export async function runPreDeployTask(context: appservice.IDeployContext, siteConfig: WebSiteManagementModels.SiteConfigResource): Promise<void> {
     await appservice.runPreDeployTask(context, context.originalDeployFsPath, siteConfig.scmType);
 
     // only executable jar or war files generated by `preDeployTask`can be deployed to java runtime appservice, while
-    // the `context.effectiveDeployFsPath` is the path of the pom file/module folder, which can not be deployed directly to Azure.
+    // the `context.effectiveDeployFsPath` points the path of the pom file/module folder when deploying a maven module.
+    // we need to modify `context.effectiveDeployFsPath` to point the executable jar or war files, of which the path
+    // can only be determined after preDeployTask is run.
     if (javaUtils.isJavaRuntime(siteConfig)) {
-        // modify `context.effectiveDeployFsPath` to the executable jar or war files,
-        // of which the path can only be determined after running preDeployTask
         const fileExtensions = await javaUtils.getJavaFileExtensions(siteConfig);
 
         // decide module path
@@ -27,7 +28,7 @@ export async function runPreDeployTask(context: appservice.IDeployContext, siteC
         const moduleFolder = isPomFile ? path.dirname(effectivePath) : effectivePath;
 
         // search for executable jar/war files under `${moduleFolder}/target`
-        const executableArtifacts = await javaUtils.getExecutableArtifactsByExtensions(path.join(moduleFolder, 'target'), fileExtensions);
+        const executableArtifacts = await tryGetDeployableArtifacts(path.join(moduleFolder, 'target'), fileExtensions);
         context.effectiveDeployFsPath = executableArtifacts?.[0] ?? '';
         // show warning if can not find executable jar/war artifacts in module.
         const message = localize('errorArtifactsNotFoundJava', 'No {0} artifacts found in maven module "{1}"', `${fileExtensions}`, moduleFolder);
